MSBuildProjectHandler: ignore optional migrations in md-tool
When migrating a project where migration is optional, do not
throw an exception if projectLoadMonitor is null, which will
happen when a project is opened through md-tool (i.e. MD
UI is not actually running). In this case just ignore migration
since it's not required. No reason for it to be fatal.